Update plugins to not make file changes if no changes are made

+++ _plugins/fix_urls.rb
@@ -9,10 +9,14 @@ module Jekyll
 
       Jekyll.logger.info "  Fixing up URLs in swagger files"
       Dir.glob(%w[./docker-hub/api/*.yaml ./engine/api/*.yaml]) do |file_name|
-        Jekyll.logger.info "    #{file_name}"
         text = File.read(file_name)
         replace = text.gsub("https://docs.docker.com", "")
-        File.open(file_name, "w") { |file| file.puts replace }
+        if text == replace
+          Jekyll.logger.info "    #{file_name} (skipped - no change)"
+        else
+          Jekyll.logger.info "    #{file_name}"
+          File.open(file_name, "w") { |file| file.puts replace }
+        end
       end
 
       end_time = Time.now
